Avanzo Ejercicio13

+++ src/main/java/com/jorgerubira/ejerciciosspringweb/repositories/FacturaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.jorgerubira.ejerciciosspringweb.entities.Factura;
 
-public interface FacturaRepository extends JpaRepository<Factura, Integer>{
+public interface FacturaRepository extends JpaRepository<Factura, Long>{
 
 }
